[move]아이콘 컴포넌트, utils 함수 폴더 이동
