+++ src/app/(auth)/_components/ResetPassword.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useAuthStore } from "@/store/authStore";
-import { passwordReg } from "@/utils/authValidation";
+import { passwordReg } from "@/lib/utils/auth/authValidation";
 import { useRouter } from "next/navigation";
 import { useEffect, useRef, useState } from "react";
 import { FaRegEyeSlash } from "react-icons/fa";
