+++ src/app/(non-header)/my-page/_components/EditNickname.tsx
@@ -3,7 +3,7 @@
 import { useUserData } from "@/hooks/useUserData";
 import { useAuthStore } from "@/store/authStore";
 import { Auth } from "@/types/auth.type";
-import { nicknameReg } from "@/utils/authValidation";
+import { nicknameReg } from "@/lib/utils/auth/authValidation";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
 import React, { useEffect, useRef } from "react";
