+++ src/app/api/chat/assistant/[id]/route.ts
@@ -1,7 +1,7 @@
 import { extractTodoItems, handleSaveChatTodo } from "@/app/api/lib/chatTodoItemUtils";
 import { formatTodoList, getTodoRequestType } from "@/app/api/lib/todoPatterns";
 import { CHAT_SESSIONS } from "@/lib/tableNames";
-import openai from "@/lib/utils/openaiClient";
+import openai from "@/lib/utils/chat/openaiClient";
 import { Chat, ChatSession, Message, MessageWithSaveButton } from "@/types/chat.session.type";
 import { Json } from "@/types/supabase";
 import { createClient } from "@/utils/supabase/server";
