+++ src/lib/utils/auth/authValidation.ts
