+++ src/lib/utils/auth/fetchUserData.ts
@@ -1,4 +1,4 @@
-import { createClient } from "./supabase/client";
+import { createClient } from "../../../utils/supabase/client";
 
 export const fetchUserData = async () => {
   const supabase = createClient();
