+++ src/lib/utils/chat/openaiClient.ts
