+++ src/lib/utils/todos/updateTodoAddress.ts
@@ -1,7 +1,5 @@
 import { UpdateTOdoAddressType } from "@/types/diary.type";
-import { supabase } from "./supabase/client";
-
-
+import { supabase } from "../../../utils/supabase/client";
 
 export const updateTodoAddress = async ({ todoId, lat, lng }: UpdateTOdoAddressType): Promise<void> => {
   const { data, error } = await supabase.from("todos").update({ address: { lat, lng } }).eq("todo_id", todoId);
